requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
flask-cors==4.0.0
requests==2.31.0
numpy==1.24.3
sympy==1.12
matplotlib==3.7.2
plotly==5.15.0
pandas==2.0.3
scipy==1.11.1
psutil==5.9.5
watchdog==3.0.0
python-telegram-bot==20.7
torch==2.1.0
sentence-transformers==2.2.2
